requirements.txt
langchain~=0.3.0
langchain-core~=0.3.0
langchain-openai~=0.2.0
langgraph~=0.2.0
openai
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
streamlit>=1.28.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0